[db] Move initialization of new db into its own file (db_init.c)
